Split the Shortcuts pane along its seams
At 304 lines it was nearly double the next-largest Settings pane and the only
one carrying a private view of its own, which is the signal that it had
outgrown the shape every other pane holds.

ShortcutRow moves to Components, where every other row-shaped view already
lives. Turning clicks into a binding moves out of KeyboardMap into
ShortcutAssignment: KeyboardMap is about what to draw, and what a sequence of
clicks means is a different question that was renting space there.

Deliberately not a reducer. ADR-0005 says one is added only where an enum state
machine with real races already exists, and that Settings panes stay plain MV.
There are no races here and no async — a reducer would be the ceremony that ADR
was written to prevent.

Moving the picking rule out did make it testable, though. Clicking a held
modifier again to let it go is real behaviour that had no coverage, because it
only existed inside a view.

No behaviour changes, so the existing tests are the proof the moves were safe.
