+++ Talkify/Input/KeyboardMap.swift
@@ -117,48 +117,6 @@ enum KeyboardMap {
     return "?"
   }
 
-  /// The four modifiers a binding can *require*. fn is missing on purpose: a
-  /// binding stores its required modifiers as `CGEventFlags` command, option,
-  /// control and shift, so fn can only ever be the bound key itself.
-  static let combiningModifiers: Set<Int64> = [55, 54, 58, 61, 59, 62, 56, 60]
-
-  /// Builds a binding from keys clicked on the drawn keyboard, in click order.
-  ///
-  /// The bound key is the first thing clicked that is not one of the four
-  /// combining modifiers — fn, a letter, Escape — and everything else becomes a
-  /// required modifier. When only combining modifiers were clicked, the last
-  /// one is the key, which is how a bare ⌥ gets bound.
-  ///
-  /// This is what lets someone assign a combination they cannot type into the
-  /// recorder, because the system or another app swallows it first.
-  static func binding(forClicked clicked: [Int64], layout: KeyboardLayout) -> KeyBinding? {
-    guard let last = clicked.last else { return nil }
-    let keyCode = clicked.first { !combiningModifiers.contains($0) } ?? last
-
-    var flags: CGEventFlags = []
-    for modifier in clicked where modifier != keyCode {
-      flags.insert(KeyBinding.modifierMask(forKeyCode: modifier))
-    }
-    // fn cannot be a required modifier, only a bound key.
-    flags.remove(.maskSecondaryFn)
-
-    let isModifierKey = KeyBinding.modifierKeyName(forKeyCode: keyCode) != nil
-    var caps: [String] = []
-    if flags.contains(.maskControl) { caps.append("⌃") }
-    if flags.contains(.maskAlternate) { caps.append("⌥") }
-    if flags.contains(.maskShift) { caps.append("⇧") }
-    if flags.contains(.maskCommand) { caps.append("⌘") }
-    caps.append(cap(for: keyCode, layout: layout))
-
-    return KeyBinding(
-      keyCode: keyCode,
-      modifierFlags: flags.rawValue,
-      isModifierKey: isModifierKey,
-      label: caps.joined(separator: " "),
-      keyEquivalent: isModifierKey ? "" : (layout.legend(for: keyCode)?.lowercased() ?? "")
-    )
-  }
-
   /// The keys a binding lights up: its own key plus whichever modifier keys it
   /// requires. A modifier appears on both sides of the board, so both light —
   /// the binding does not care which one is pressed unless it named a side.
